Add endpoints to base & format
