+++ src/core/agent/base.py
@@ -1,9 +1,13 @@
 import abc
+import logging
 from abc import ABC
-from typing import List, Optional
+from typing import List, Optional, Dict, Any
 
 from langchain.agents import Tool
-from steamship.invocable import PackageService
+from steamship import SteamshipError, Block
+from steamship.invocable import PackageService, post
+
+from utils import is_valid_uuid
 
 
 def response_for_exception(e: Optional[Exception]) -> str:
@@ -43,3 +47,32 @@ class BaseAgent(PackageService, ABC):
 
     def run(self, prompt: str) -> str:
         return self.get_agent().run(input=prompt)
+
+    @post("answer", public=True)
+    def answer(
+        self, question: str, chat_session_id: Optional[str] = None
+    ) -> List[Dict[str, Any]]:
+        """Endpoint that implements the contract for Steamship embeddable chat widgets.
+        This is a PUBLIC endpoint since these webhooks do not pass a token."""
+        logging.info(f"/answer: {question} {chat_session_id}")
+
+        try:
+            response = self.run(question)
+        except SteamshipError as e:
+            response = [response_for_exception(e)]
+
+        answer = []
+        for part_response in response if isinstance(response, list) else [response]:
+            if is_valid_uuid(part_response):
+                block = Block.get(self.client, _id=part_response).dict()
+                block["who"] = "bot"
+                answer.append(block)
+            else:
+                answer.append({"message": part_response, "who": "bot"})
+
+        return answer
+
+    @post("info")
+    def info(self) -> dict:
+        """Endpoint returning information about this bot."""
+        return {"telegram": "Hello There!"}
