+++ src/core/agent/chat_agent.py
@@ -14,9 +14,13 @@ class ChatAgent(BaseAgent):
     name: str = "MyAgent"
 
     def get_agent(self):
-        llm = OpenAIChat(client=self.client, temperature=TEMPERATURE, model_name=MODEL_NAME)
+        llm = OpenAIChat(
+            client=self.client, temperature=TEMPERATURE, model_name=MODEL_NAME
+        )
 
-        sys_message = f"{PREFIX}\n Please take on the personality of {self.get_personality()}"
+        sys_message = (
+            f"{PREFIX}\n Please take on the personality of {self.get_personality()}"
+        )
 
         tools = self.get_tools()
         memory = ConversationBufferWindowMemory(
@@ -25,13 +29,11 @@ class ChatAgent(BaseAgent):
         )
 
         agent = ConversationalChatAgent.from_llm_and_tools(
-            llm, tools,
-            system_message=sys_message,
-            output_parser=ChatCustomParser()
+            llm, tools, system_message=sys_message, output_parser=ChatCustomParser()
         )
         return AgentExecutor.from_agent_and_tools(
             agent=agent,
             tools=tools,
             memory=memory,
-            verbose=self.is_verbose_logging_enabled()
+            verbose=self.is_verbose_logging_enabled(),
         )
