+++ src/core/chat_parser.py
@@ -1,6 +1,8 @@
 from typing import Any
 
-from langchain.agents.conversational_chat.base import AgentOutputParser as BaseChatAgentOutputParser
+from langchain.agents.conversational_chat.base import (
+    AgentOutputParser as BaseChatAgentOutputParser,
+)
 from langchain.agents.conversational_chat.output_parser import ConvoOutputParser
 
 
@@ -12,5 +14,5 @@ class ChatCustomParser(ConvoOutputParser):
     def parse(self, text: str) -> Any:
         cleaned_output = text.strip()
         if cleaned_output.startswith("AI: "):
-            cleaned_output = cleaned_output[len("AI: "):]
+            cleaned_output = cleaned_output[len("AI: ") :]
         return super().parse(cleaned_output)
